% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_rank_frequency}
\alias{plot_rank_frequency}
\title{Plot Rank-Frequency Distribution (Zipf Plot)}
\usage{
plot_rank_frequency(freq_tbl, top = 5000, name_col = "word", p_col = "p")
}
\arguments{
\item{freq_tbl}{A frequency table with relative frequency column.}

\item{top}{Integer. Number of top terms to include in plot. Default is 5000.}

\item{name_col}{Character. Name of column containing term names.
Default is "word".}

\item{p_col}{Character. Name of column containing relative frequencies.
Default is "p".}
}
\value{
A ggplot2 object showing log-log rank-frequency plot.
}
\description{
This function creates a log-log plot showing the relationship between
rank and frequency, which typically follows Zipf's law in natural language.
}
\details{
Zipf's law predicts that the frequency of a word is inversely proportional
to its rank. On a log-log plot, this appears as a straight line with
negative slope around -1.
}
\examples{
\dontrun{
# Plot Zipf distribution for words
corpus <- load_corpus("en_US")
unigrams <- tokenize_unigrams(corpus)
freq_table <- freq_unigrams(unigrams)

plot_rank_frequency(freq_table, top = 1000)

# Compare bigram distribution
bigrams <- tokenize_bigrams(corpus)
freq_bi <- freq_bigrams(bigrams)
plot_rank_frequency(freq_bi, top = 2000)
}

}
