% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_top_terms}
\alias{plot_top_terms}
\title{Plot Top Terms by Frequency}
\usage{
plot_top_terms(freq_tbl, top = 20, name_col = "word")
}
\arguments{
\item{freq_tbl}{A frequency table with count column and term identification.}

\item{top}{Integer. Number of top terms to display. Default is 20.}

\item{name_col}{Character. Name of column containing term names.
Default is "word". If missing, will attempt to create from w1..wn columns.}
}
\value{
A ggplot2 object showing horizontal bars for top terms.
}
\description{
This function creates a horizontal bar chart showing the most frequent
terms in the corpus. Useful for quick inspection of dominant vocabulary.
}
\examples{
\dontrun{
# Plot top words
corpus <- load_corpus("en_US")
unigrams <- tokenize_unigrams(corpus)
freq_table <- freq_unigrams(unigrams)

plot_top_terms(freq_table, top = 15)

# Plot top bigrams
bigrams <- tokenize_bigrams(corpus)
freq_bi <- freq_bigrams(bigrams)
plot_top_terms(freq_bi, top = 10)
}

}
