% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenization.R
\name{tokenize_bigrams}
\alias{tokenize_bigrams}
\title{Create Bigrams (2-grams)}
\usage{
tokenize_bigrams(corpus, text_col = "text")
}
\arguments{
\item{corpus}{A tibble containing text data.}

\item{text_col}{Character or unquoted column name containing text. Default is "text".}
}
\value{
A tibble with bigrams in the "ng" column.
}
\description{
Convenience function to create bigrams (2-word sequences) from text.
This is equivalent to \code{tokenize_ngrams(corpus, n = 2)}.
}
\examples{
\dontrun{
corpus <- load_corpus("en_US")
bigrams <- tokenize_bigrams(corpus)
head(bigrams)
}

}
