% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenization.R
\name{tokenize_trigrams}
\alias{tokenize_trigrams}
\title{Create Trigrams (3-grams)}
\usage{
tokenize_trigrams(corpus, text_col = "text")
}
\arguments{
\item{corpus}{A tibble containing text data.}

\item{text_col}{Character or unquoted column name containing text. Default is "text".}
}
\value{
A tibble with trigrams in the "ng" column.
}
\description{
Convenience function to create trigrams (3-word sequences) from text.
This is equivalent to \code{tokenize_ngrams(corpus, n = 3)}.
}
\examples{
\dontrun{
corpus <- load_corpus("en_US")
trigrams <- tokenize_trigrams(corpus)
head(trigrams)
}

}
